Address comments and move all files related to block cache analyzer into the block cache analyzer directory.
