+++ .gitignore
@@ -34,7 +34,7 @@ manifest_dump
 sst_dump
 blob_dump
 block_cache_trace_analyzer
-tools/*pyc
+tools/block_cache_analyzer/*.pyc
 column_aware_encoding_exp
 util/build_version.cc
 build_tools/VALGRIND_LOGS/
