+++ tools/block_cache_analyzer/block_cache_trace_analyzer.cc
@@ -5,7 +5,7 @@
 
 #ifndef ROCKSDB_LITE
 #ifdef GFLAGS
-#include "tools/block_cache_trace_analyzer.h"
+#include "tools/block_cache_analyzer/block_cache_trace_analyzer.h"
 
 #include <algorithm>
 #include <cinttypes>
@@ -1395,13 +1395,12 @@ Status BlockCacheTraceAnalyzer::WriteHumanReadableTraceRecord(
   }
   int ret = snprintf(
       trace_record_buffer_, sizeof(trace_record_buffer_),
-      "%" PRIu64 ",%" PRIu64 ",%u,%" PRIu64 ",%" PRIu64 ",%" PRIu32 ",%" PRIu64
-      ""
-      ",%u,%u,%" PRIu64 ",%" PRIu64 ",%" PRIu64 ",%u\n",
+      "%" PRIu64 ",%" PRIu64 ",%u,%" PRIu64 ",%" PRIu64 ",%s,%" PRIu32
+      ",%" PRIu64 ",%u,%u,%" PRIu64 ",%" PRIu64 ",%" PRIu64 ",%u\n",
       access.access_timestamp, block_id, access.block_type, access.block_size,
-      access.cf_id, access.level, access.sst_fd_number, access.caller,
-      access.no_insert, access.get_id, get_key_id, access.referenced_data_size,
-      access.is_cache_hit);
+      access.cf_id, access.cf_name.c_str(), access.level, access.sst_fd_number,
+      access.caller, access.no_insert, access.get_id, get_key_id,
+      access.referenced_data_size, access.is_cache_hit);
   if (ret < 0) {
     return Status::IOError("failed to format the output");
   }
@@ -2134,6 +2133,7 @@ int block_cache_trace_analyzer_tool(int argc, char** argv) {
         analyzer.WriteAccessTimeline(label, kSecondInHour, false);
       } else {
         analyzer.WriteAccessTimeline(label, kSecondInMinute, false);
+        analyzer.WriteAccessTimeline(label, kSecondInHour, false);
       }
     }
   }
