+++ tools/block_cache_analyzer/block_cache_trace_analyzer.h
